+++ src/ska_mid_dish_manager/component_managers/dish_manager_cm.py
@@ -204,7 +204,7 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
             lastcommandedmode=("0.0", ""),
             lastcommandinvoked=("0.0", ""),
             lastcommandfailure=("0.0", "", ""),
-            lasttangoerror=("0.0", "", "", ""),
+            lasttangoeventerror=("0.0", "", "", ""),
             dscctrlstate=DscCtrlState.NO_AUTHORITY,
             rfcmplllock=B5dcPllState.NOT_LOCKED,
             rfcmhattenuation=0.0,
@@ -930,20 +930,14 @@ class DishManagerComponentManager(TaskExecutorComponentManager):
         """Callback triggered by the component manager of the subservient device
         when an error event is received on one of its attribute subscriptions.
 
-        DishManager reports the details of the error on the lastTangoError attribute.
+        DishManager reports the details of the error on the lastTangoEventError attribute.
 
         :param device_name: TRL of the device which emitted the error event
         :param attribute_name: name of the attribute the error event was emitted for
         :param reason: the reason reported on the tango error
         """
-        self.logger.debug(
-            "Tango error reported by %s on attribute %s: %s",
-            device_name,
-            attribute_name,
-            reason,
-        )
         self._update_component_state(
-            lasttangoerror=(str(time.time()), device_name, attribute_name, reason)
+            lasttangoeventerror=(str(time.time()), device_name, attribute_name, reason)
         )
 
     def _update_connection_state_attribute(